requirements_linux.txt
ruamel.yaml
opencv-python
matplotlib
pandas
scipy
gym
docopt
mlagents_envs==0.10.0
https://download.pytorch.org/whl/cpu/torch-1.4.0%2Bcpu-cp36-cp36m-linux_x86_64.whl
https://download.pytorch.org/whl/cpu/torchvision-0.5.0%2Bcpu-cp36-cp36m-linux_x86_64.whl
tensorboard==2.0.2
gym-minigrid==1.0.1
procgen
